0ColorBlindnessAssistant
Initial commit. An extension whose main purpose is to serve as an assistant to users with color vision deficiency (CVD). It provides compensation effects (it allows users with CVD to better differentiate colors that they can't see), simulation effects (it allows developers to see how users with CVD see certain combination of colors and ascertain if a color palette is *CVD friendly*) and a *color naming tool* that allows to get the name of a color on the screen.


TODO:

- Complete Spanish localization.
- Complete help page.
- Complete contributors section.
- Implement a wizard-like mechanism to apply effects.